@@ -1,3 +1,32 @@
-module.exports = {
+export default {
   presets: ['module:@react-native/babel-preset'],
+  plugins: [
+    [
+      'module-resolver',
+      {
+        root: ['./src'],
+        extensions: [
+          '.ios.ts',
+          '.android.ts',
+          '.ts',
+          '.ios.tsx',
+          '.android.tsx',
+          '.tsx',
+          '.jsx',
+          '.js',
+          '.json',
+        ],
+        alias: {
+          '@components': './src/components',
+          '@screens': './src/screens',
+          '@navigation': './src/navigation',
+          '@assets': './src/assets',
+          '@hooks': './src/hooks',
+          '@utils': './src/utils',
+          '@types': './src/types',
+          '@config': './src/config',
+        },
+      },
+    ],
+  ],
 };
