@@ -1,7 +1,10 @@
-module.exports = {
+export default {
   arrowParens: 'avoid',
   bracketSameLine: true,
   bracketSpacing: false,
   singleQuote: true,
   trailingComma: 'all',
+  semi: true,
+  tabWidth: 2,
+  useTabs: false,
 };
